refactor(rust): extract shared sceneworks-core util modules (sc-1640/1641/1642/1643/1644/1661)
Consolidate copy-pasted helpers that the full-repo review (epic 1628)
flagged as duplicated across the Rust stores, worker, and API:

- time.rs: utc_now/now_unix_seconds/format_unix_seconds (+civil math),
  dedupes 4 byte-identical store copies (sc-1640).
- slug.rs: one shared slugify; worker drops its byte-identical copy
  (sc-1641). The API _-separator id variants genuinely differ and stay.
- store_util.rs: read_json/write_json, relative_string, path-safety
  predicates, optional_* accessors, random_hex, parse_string_enum,
  shared by project/character/training/jobs stores (sc-1642).
- asset_index.rs: AssetRecord + row mapper, asset_sidecars/collect_sidecars,
  normalize_asset, the assets upsert SQL, and a single ASSET_FOLDERS
  (sc-1643). The shared ASSET_FOLDERS includes assets/documents, so
  character reindex no longer silently drops document sidecars (sc-1644).
- worker/API now_rfc3339 route through format_unix_seconds(now_unix_seconds()),
  removing the .expect() panic surface and the sole `time` crate dependency
  in both crates (sc-1661).

Net ~570 fewer lines. `npm run rust:check` (fmt, clippy -D warnings,
test, build) passes.
