+++ crates/sceneworks-core/src/jobs_store.rs
@@ -1,6 +1,5 @@
 use std::fs;
 use std::path::{Path, PathBuf};
-use std::time::{SystemTime, UNIX_EPOCH};
 
 use parking_lot::Mutex;
 use rusqlite::{params, params_from_iter, Connection, OptionalExtension, Row};
@@ -11,6 +10,8 @@ use crate::contracts::{
     ContractNumber, JobSnapshot, JobStatus, JobType, ProgressStage, QueueSummary, WorkerCapability,
     WorkerSnapshot, WorkerStatus, WorkerUtilizationSnapshot,
 };
+use crate::store_util::parse_string_enum;
+use crate::time::{format_unix_seconds, now_unix_seconds, utc_now};
 
 pub const ACTIVE_STATUSES: &[&str] = &[
     "preparing",
@@ -1060,14 +1061,6 @@ where
     value.and_then(|text| serde_json::from_str::<T>(text).ok())
 }
 
-fn parse_string_enum<T>(value: &str) -> T
-where
-    T: DeserializeOwned,
-{
-    serde_json::from_value(Value::String(value.to_owned()))
-        .expect("string enum deserialization is infallible")
-}
-
 fn number_from_f64(value: f64) -> ContractNumber {
     Number::from_f64(value).unwrap_or_else(|| Number::from(0))
 }
@@ -1077,27 +1070,6 @@ fn elapsed_seconds(started_at: &str, completed_at: Option<&str>) -> Option<Contr
     let ended = completed_at.map_or_else(|| Some(now_unix_seconds()), parse_utc_seconds)?;
     let seconds = ended.saturating_sub(started).max(0);
     Some(Number::from(seconds))
-}
-
-fn utc_now() -> String {
-    format_unix_seconds(now_unix_seconds())
-}
-
-fn now_unix_seconds() -> i64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|duration| i64::try_from(duration.as_secs()).unwrap_or(i64::MAX))
-        .unwrap_or(0)
-}
-
-fn format_unix_seconds(timestamp: i64) -> String {
-    let days = timestamp.div_euclid(86_400);
-    let seconds_of_day = timestamp.rem_euclid(86_400);
-    let (year, month, day) = civil_from_days(days);
-    let hour = seconds_of_day / 3_600;
-    let minute = (seconds_of_day % 3_600) / 60;
-    let second = seconds_of_day % 60;
-    format!("{year:04}-{month:02}-{day:02}T{hour:02}:{minute:02}:{second:02}Z")
 }
 
 fn parse_utc_seconds(value: &str) -> Option<i64> {
@@ -1138,21 +1110,6 @@ fn parse_utc_seconds(value: &str) -> Option<i64> {
         }
     }
     Some(days_from_civil(year, month, day) * 86_400 + hour * 3_600 + minute * 60 + second)
-}
-
-fn civil_from_days(days: i64) -> (i64, i64, i64) {
-    let adjusted_days = days + 719_468;
-    let era = adjusted_days.div_euclid(146_097);
-    let day_of_era = adjusted_days - era * 146_097;
-    let year_of_era =
-        (day_of_era - day_of_era / 1_460 + day_of_era / 36_524 - day_of_era / 146_096) / 365;
-    let mut year = year_of_era + era * 400;
-    let day_of_year = day_of_era - (365 * year_of_era + year_of_era / 4 - year_of_era / 100);
-    let month_prime = (5 * day_of_year + 2) / 153;
-    let day = day_of_year - (153 * month_prime + 2) / 5 + 1;
-    let month = month_prime + if month_prime < 10 { 3 } else { -9 };
-    year += i64::from(month <= 2);
-    (year, month, day)
 }
 
 fn days_from_civil(year: i32, month: u32, day: u32) -> i64 {
