+++ crates/sceneworks-core/src/lib.rs
@@ -1,10 +1,14 @@
 pub mod app_paths;
+pub mod asset_index;
 pub mod character_store;
 pub mod contracts;
 pub mod jobs_store;
 pub mod lora_family;
 pub mod lora_url;
 pub mod project_store;
+pub mod slug;
+pub mod store_util;
+pub mod time;
 pub mod training;
 pub mod training_store;
