+++ crates/sceneworks-core/src/project_store.rs
@@ -1,12 +1,14 @@
 use std::fs;
 use std::path::{Path, PathBuf};
-use std::time::{SystemTime, UNIX_EPOCH};
 
 use parking_lot::Mutex;
-use rusqlite::{params, Connection, OptionalExtension, Row};
+use rusqlite::{params, Connection, OptionalExtension};
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Map, Value};
 
+use crate::asset_index::{
+    asset_sidecars, normalize_asset, row_to_asset_record, upsert_asset_row, AssetRecord,
+};
 use crate::character_store::{
     apply_character_migrations, clear_character_tables, reindex_characters_on_connection,
     CharacterStore,
@@ -16,6 +18,12 @@ pub use crate::character_store::{
     CharacterLoraUpdateInput, CharacterMutationResult, CharacterReferenceInput,
     CharacterReferenceUpdateInput, CharacterUpdateInput, CHARACTER_SIDECAR_PATTERN,
 };
+use crate::slug::slugify;
+use crate::store_util::{
+    is_safe_relative_path, optional_f64, optional_str, optional_u64, random_hex, read_json,
+    relative_string, write_json,
+};
+use crate::time::utc_now;
 use crate::training::TrainingDataset;
 use crate::training_store::{
     apply_training_dataset_migrations, TrainingCaptionSidecarsResult,
@@ -24,7 +32,6 @@ use crate::training_store::{
     TrainingDatasetSummary, TrainingDatasetUpdateInput,
 };
 
-pub const ASSET_SIDECAR_PATTERN: &str = "*.sceneworks.json";
 pub const PROJECT_FOLDERS: &[&str] = &[
     "assets/images",
     "assets/videos",
@@ -42,16 +49,6 @@ pub const PROJECT_FOLDERS: &[&str] = &[
     "trash",
     "cache",
 ];
-const ASSET_FOLDERS: &[&str] = &[
-    "assets/images",
-    "assets/videos",
-    "assets/uploads",
-    "assets/frames",
-    "assets/renders",
-    "assets/documents",
-    "trash",
-];
-
 pub type ProjectStoreResult<T> = Result<T, ProjectStoreError>;
 
 #[derive(Debug)]
@@ -1273,12 +1270,6 @@ struct ReindexCounts {
     timelines: u32,
 }
 
-#[derive(Debug)]
-struct AssetRecord {
-    file_path: Option<String>,
-    sidecar_path: Option<String>,
-}
-
 pub fn apply_project_migrations(connection: &Connection) -> ProjectStoreResult<()> {
     connection.execute_batch(
         "
@@ -2011,34 +2002,7 @@ fn index_asset_on_connection(
     asset: &Value,
     sidecar_rel: Option<&str>,
 ) -> ProjectStoreResult<()> {
-    let status = asset.get("status").unwrap_or(&Value::Null);
-    connection.execute(
-        "
-        insert or replace into assets (
-          id, type, display_name, file_path, generation_set_id, created_at,
-          favorite, rating, rejected, trashed, sidecar_path
-        ) values (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11)
-        ",
-        params![
-            required_str(asset, "id")?,
-            required_str(asset, "type")?,
-            required_str(asset, "displayName")?,
-            asset
-                .get("file")
-                .and_then(|file| optional_str(file, "path"))
-                .ok_or_else(|| ProjectStoreError::BadRequest(
-                    "Asset file path is required".to_owned()
-                ))?,
-            optional_str(asset, "generationSetId"),
-            required_str(asset, "createdAt")?,
-            optional_bool(status, "favorite").unwrap_or(false),
-            optional_u64(status, "rating").unwrap_or(0),
-            optional_bool(status, "rejected").unwrap_or(false),
-            optional_bool(status, "trashed").unwrap_or(false),
-            sidecar_rel,
-        ],
-    )?;
-    Ok(())
+    upsert_asset_row(connection, asset, sidecar_rel)
 }
 
 fn find_asset_record(
@@ -2090,13 +2054,6 @@ fn find_asset_sidecar_path(
     Ok(None)
 }
 
-fn row_to_asset_record(row: &Row<'_>) -> rusqlite::Result<AssetRecord> {
-    Ok(AssetRecord {
-        file_path: row.get(0)?,
-        sidecar_path: row.get(1)?,
-    })
-}
-
 fn purge_asset_record(project_path: &Path, asset_id: &str) -> ProjectStoreResult<()> {
     let mut connection = connect_project_db(project_path)?;
     let transaction = connection.transaction()?;
@@ -2108,60 +2065,6 @@ fn purge_asset_record(project_path: &Path, asset_id: &str) -> ProjectStoreResult
 
 fn project_has_sidecars(project_path: &Path) -> bool {
     asset_sidecars(project_path).is_ok_and(|paths| !paths.is_empty())
-}
-
-fn asset_sidecars(project_path: &Path) -> ProjectStoreResult<Vec<PathBuf>> {
-    let mut sidecars = Vec::new();
-    for folder in ASSET_FOLDERS {
-        collect_sidecars(&project_path.join(folder), &mut sidecars)?;
-    }
-    let timeline_dir = project_path.join("timelines");
-    sidecars.retain(|path| !path.starts_with(&timeline_dir));
-    Ok(sidecars)
-}
-
-fn collect_sidecars(path: &Path, sidecars: &mut Vec<PathBuf>) -> ProjectStoreResult<()> {
-    if !path.exists() {
-        return Ok(());
-    }
-    for entry in fs::read_dir(path)? {
-        let entry = entry?;
-        let path = entry.path();
-        if path.is_dir() {
-            collect_sidecars(&path, sidecars)?;
-        } else if path
-            .file_name()
-            .and_then(|value| value.to_str())
-            .is_some_and(|name| name.ends_with(".sceneworks.json"))
-        {
-            sidecars.push(path);
-        }
-    }
-    Ok(())
-}
-
-fn normalize_asset(
-    project_id: &str,
-    project_path: &Path,
-    sidecar_path: &Path,
-) -> ProjectStoreResult<Value> {
-    let mut asset = read_json(sidecar_path)?;
-    if let Some(path) = asset.pointer("/file/path").and_then(Value::as_str) {
-        let normalized_path = path.replace('\\', "/");
-        if let Some(object) = asset.as_object_mut() {
-            object.insert(
-                "url".to_owned(),
-                Value::String(format!(
-                    "/api/v1/projects/{project_id}/files/{normalized_path}"
-                )),
-            );
-        }
-    }
-    let sidecar_rel = relative_string(project_path, sidecar_path)?;
-    if let Some(object) = asset.as_object_mut() {
-        object.insert("sidecarPath".to_owned(), Value::String(sidecar_rel));
-    }
-    Ok(asset)
 }
 
 fn normalize_person_track(project_path: &Path, track_path: &Path) -> ProjectStoreResult<Value> {
@@ -2276,28 +2179,6 @@ fn validate_normalized_box(value: &Value) -> ProjectStoreResult<Value> {
     Ok(Value::Object(box_out))
 }
 
-fn read_json(path: &Path) -> ProjectStoreResult<Value> {
-    let payload = fs::read_to_string(path)?;
-    Ok(serde_json::from_str(&payload)?)
-}
-
-fn write_json(path: &Path, payload: &Value) -> ProjectStoreResult<()> {
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent)?;
-    }
-    let mut output = serde_json::to_string_pretty(payload)?;
-    output.push('\n');
-    let tmp_path = path.with_extension(format!(
-        "{}.tmp",
-        path.extension()
-            .and_then(|extension| extension.to_str())
-            .unwrap_or("json")
-    ));
-    fs::write(&tmp_path, output)?;
-    fs::rename(tmp_path, path)?;
-    Ok(())
-}
-
 fn read_dir_paths(path: &Path) -> ProjectStoreResult<Vec<PathBuf>> {
     if !path.exists() {
         return Ok(Vec::new());
@@ -2305,22 +2186,6 @@ fn read_dir_paths(path: &Path) -> ProjectStoreResult<Vec<PathBuf>> {
     fs::read_dir(path)?
         .map(|entry| entry.map(|entry| entry.path()).map_err(Into::into))
         .collect()
-}
-
-fn relative_string(root: &Path, path: &Path) -> ProjectStoreResult<String> {
-    Ok(path
-        .strip_prefix(root)
-        .map_err(|_| ProjectStoreError::BadRequest("Path is outside project".to_owned()))?
-        .to_string_lossy()
-        .replace('\\', "/"))
-}
-
-fn is_safe_relative_path(relative_path: &str) -> bool {
-    !relative_path.trim().is_empty()
-        && !relative_path.contains('\\')
-        && Path::new(relative_path)
-            .components()
-            .all(|component| matches!(component, std::path::Component::Normal(_)))
 }
 
 fn is_safe_track_id(track_id: &str) -> bool {
@@ -2333,46 +2198,6 @@ fn is_safe_track_id(track_id: &str) -> bool {
 fn required_str<'a>(payload: &'a Value, key: &str) -> ProjectStoreResult<&'a str> {
     optional_str(payload, key)
         .ok_or_else(|| ProjectStoreError::BadRequest(format!("Missing required field: {key}")))
-}
-
-fn optional_str<'a>(payload: &'a Value, key: &str) -> Option<&'a str> {
-    payload.get(key).and_then(Value::as_str)
-}
-
-fn optional_bool(payload: &Value, key: &str) -> Option<bool> {
-    payload.get(key).and_then(Value::as_bool)
-}
-
-fn optional_u64(payload: &Value, key: &str) -> Option<u64> {
-    payload.get(key).and_then(Value::as_u64)
-}
-
-fn optional_f64(payload: &Value, key: &str) -> Option<f64> {
-    payload.get(key).and_then(Value::as_f64)
-}
-
-fn slugify(value: &str, fallback: &str, max_length: Option<usize>) -> String {
-    let mut slug = String::new();
-    let mut previous_dash = false;
-    for character in value.trim().chars() {
-        if character.is_ascii_alphanumeric() {
-            slug.push(character.to_ascii_lowercase());
-            previous_dash = false;
-        } else if !previous_dash && !slug.is_empty() {
-            slug.push('-');
-            previous_dash = true;
-        }
-    }
-    while slug.ends_with('-') {
-        slug.pop();
-    }
-    if slug.is_empty() {
-        slug = fallback.to_owned();
-    }
-    if let Some(max_length) = max_length {
-        slug.truncate(max_length);
-    }
-    slug
 }
 
 fn safe_filename(value: &str, fallback: &str) -> String {
@@ -2444,51 +2269,6 @@ fn move_or_copy_file(source: &Path, destination: &Path) -> ProjectStoreResult<()
             Ok(())
         }
     }
-}
-
-fn random_hex(bytes: usize) -> ProjectStoreResult<String> {
-    let connection = Connection::open_in_memory()?;
-    Ok(connection.query_row(
-        &format!("select lower(hex(randomblob({bytes})))"),
-        [],
-        |row| row.get(0),
-    )?)
-}
-
-fn utc_now() -> String {
-    format_unix_seconds(now_unix_seconds())
-}
-
-fn now_unix_seconds() -> i64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|duration| i64::try_from(duration.as_secs()).unwrap_or(i64::MAX))
-        .unwrap_or(0)
-}
-
-fn format_unix_seconds(timestamp: i64) -> String {
-    let days = timestamp.div_euclid(86_400);
-    let seconds_of_day = timestamp.rem_euclid(86_400);
-    let (year, month, day) = civil_from_days(days);
-    let hour = seconds_of_day / 3_600;
-    let minute = (seconds_of_day % 3_600) / 60;
-    let second = seconds_of_day % 60;
-    format!("{year:04}-{month:02}-{day:02}T{hour:02}:{minute:02}:{second:02}Z")
-}
-
-fn civil_from_days(days: i64) -> (i64, i64, i64) {
-    let adjusted_days = days + 719_468;
-    let era = adjusted_days.div_euclid(146_097);
-    let day_of_era = adjusted_days - era * 146_097;
-    let year_of_era =
-        (day_of_era - day_of_era / 1_460 + day_of_era / 36_524 - day_of_era / 146_096) / 365;
-    let mut year = year_of_era + era * 400;
-    let day_of_year = day_of_era - (365 * year_of_era + year_of_era / 4 - year_of_era / 100);
-    let month_prime = (5 * day_of_year + 2) / 153;
-    let day = day_of_year - (153 * month_prime + 2) / 5 + 1;
-    let month = month_prime + if month_prime < 10 { 3 } else { -9 };
-    year += i64::from(month <= 2);
-    (year, month, day)
 }
 
 #[cfg(test)]
